Linux 64bit: Blowfish problem turned out to be SHA1 porting issue. Fixed.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@3875 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/src/core/sha1.cpp
@@ -40,13 +40,13 @@
 
 /* Hash a single 512-bit block. This is the core of the algorithm. */
 
-static void SHA1Transform(unsigned long state[5],
+static void SHA1Transform(uint32 state[5],
                           const unsigned char buffer[64])
 {
-  unsigned long a, b, c, d, e;
+  uint32 a, b, c, d, e;
   typedef union {
     unsigned char c[64];
-    unsigned long l[16];
+    uint32 l[16];
   } CHAR64LONG16;
   CHAR64LONG16* block;
 //commented, because otherwise we change the buffer
